Pagination is not part of the JSON:API specification
The JSON:API specification does not specify how pagination works. There is an example in the specification, but that is not part of the specification. See the qquote below:

```
Note: JSON:API is agnostic about the pagination strategy used by a server. Effective pagination strategies include (but are not limited to): page-based, offset-based, and cursor-based. The page query parameter can be used as a basis for any of these strategies. For example, a page-based strategy might use query parameters such as page[number] and page[size], an offset-based strategy might use page[offset] and page[limit], while a cursor-based strategy might use page[cursor].
```